pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "personabench"
version = "0.1.0"
description = "Benchmark harness for evaluating AI personas across plan-act-react tasks"
authors = [
  { name = "AI Safety Collective" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pydantic>=2.6,<3",
  "pyyaml>=6.0",
  "typing-extensions>=4.8",
  "fastapi>=0.111.0,<1.0",
  "uvicorn[standard]>=0.29.0,<1.0",
  "langchain>=0.2.0,<0.3"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff"
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "UP", "B", "S", "W"]
ignore = ["E203", "E266", "E501"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]
